Update- user creation in mongodb
